Check database version in CLI
